chore: merge basic plugin
