--- packages/compat/webpack/src/plugins/basic.ts
@@ -1,35 +1,0 @@
-import path from 'path';
-import { applyBasicPlugin, type RsbuildPlugin } from '@rsbuild/shared';
-
-/**
- * Provide some basic configs of webpack
- */
-export const pluginBasic = (): RsbuildPlugin => ({
-  name: 'rsbuild-webpack:basic',
-
-  setup(api) {
-    applyBasicPlugin(api);
-
-    api.modifyBundlerChain(async (chain, { env }) => {
-      // Disable webpack performance hints.
-      // These logs are too complex
-      chain.performance.hints(false);
-
-      // This will be futureDefaults in webpack 6
-      chain.module.parser.merge({
-        javascript: {
-          exportsPresence: 'error',
-        },
-      });
-
-      if (env === 'development') {
-        // Set correct path for source map
-        // this helps VS Code break points working correctly in monorepo
-        chain.output.devtoolModuleFilenameTemplate(
-          (info: { absoluteResourcePath: string }) =>
-            path.resolve(info.absoluteResourcePath).replace(/\\/g, '/'),
-        );
-      }
-    });
-  },
-});
