+++ packages/core/src/provider/shared.ts
@@ -10,7 +10,7 @@ import { awaitableGetter, type Plugins } from '@rsbuild/shared';
 export const applyDefaultPlugins = (plugins: Plugins) =>
   awaitableGetter<RsbuildPlugin>([
     import('./plugins/transition').then((m) => m.pluginTransition()),
-    import('./plugins/basic').then((m) => m.pluginBasic()),
+    plugins.basic(),
     plugins.entry(),
     // plugins.cache(),
     plugins.target(),
