--- packages/shared/src/apply/basic.ts
@@ -1,19 +1,0 @@
-import type { RsbuildPluginAPI } from '../types';
-import { TARGET_ID_MAP } from '../constants';
-
-export function applyBasicPlugin(api: RsbuildPluginAPI) {
-  api.modifyBundlerChain((chain, { isProd, target }) => {
-    chain.name(TARGET_ID_MAP[target]);
-    // The base directory for resolving entry points and loaders from the configuration.
-    chain.context(api.context.rootPath);
-
-    chain.mode(isProd ? 'production' : 'development');
-
-    chain.merge({
-      infrastructureLogging: {
-        // Using `error` level to avoid `cache.PackFileCacheStrategy` logs
-        level: 'error',
-      },
-    });
-  });
-}
