+++ packages/shared/src/apply/index.ts
@@ -1,6 +1,5 @@
 export * from './html';
 export * from './output';
 export * from './resolve';
-export * from './basic';
 export * from './hmr';
 export * from './babel';
